% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2freq.R
\name{fmriu.freq.bandpass_fft}
\alias{fmriu.freq.bandpass_fft}
\title{Bandpass Fourier Transform}
\usage{
fmriu.freq.bandpass_fft(signal, tr = NaN, lc = NaN, hc = NaN)
}
\arguments{
\item{signal}{[nbin, nroi]: an input signal.}

\item{tr}{[1]: the repetition time of the data (in seconds).}

\item{lc}{[1]: the lower cutoff (in Hz) for the fourier transform. fourier components below this will be set to 0.}

\item{hc}{[1]: the high cutoff (in Hz) for the fourier transform. fourier components above this will be set to 0.}
}
\value{
f [nbin/2, nroi]: the single-sided frequency domain of the signal per roi.
}
\description{
A utility function for bandpass-fourier-transforming timeseries data.
}
\author{
Eric Bridgeford
}
