% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq2div.R
\name{fmriu.freq.freq2div}
\alias{fmriu.freq.freq2div}
\title{Frequency to Divergence}
\usage{
fmriu.freq.freq2div(observations)
}
\arguments{
\item{observations}{[[n]][t, nroi] a list of n timeseries with t timesteps and nroi rois.}

\item{spectrum='amp':}{the frequency spectrum to use for computations.}
}
\value{
div_mtx [[n]][nroi, nroi] a list of n divergence matrices where each edge is a divergence between two pairs of rois.
}
\description{
a function that converts lists of frequency domain signals to divergence matrices.
}
\keyword{divergence}
\keyword{frequency,}
