% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2freq.R
\name{fmriu.freq.freq_seq}
\alias{fmriu.freq.freq_seq}
\title{Frequency per Bin}
\usage{
fmriu.freq.freq_seq(fs, nt)
}
\arguments{
\item{fs}{[1]: the sampling frequency of the signal.}

\item{nt}{[1]: the number of bins.}
}
\value{
freq_bin [nt/2]: the frequency per bin of the frequency-domain signal.
}
\description{
A utility to compute the frequency per bin of a frequency-domain signal.
}
