% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_timeseries.R
\name{fmriu.io.collection.open_timeseries}
\alias{fmriu.io.collection.open_timeseries}
\title{Open Timeseries Collection}
\usage{
fmriu.io.collection.open_timeseries(basepath, datasets = "", atlases = "",
  verbose = FALSE, rtype = "list")
}
\arguments{
\item{basepath:}{[1] a path to a properly spec'd directory. See the example in `fmriutils.html` for details.}

\item{datasets:}{[d] the datasets as a vector. Note that we expect a folder `<basepath>/<dataset>` for `<dataset> in <datasets>`.}

\item{atlases:}{[a] the atlases as a vector. Note that we expect a folder `<basepath>/<dataset>/timeseries/roi/<atlas>` for `<atlas> in <atlases>`.}

\item{verbose=FALSE:}{whether to print the id of the scan being loaded.}

\item{rtype='list':}{the type of output to return. Options are 'list' and 'array'.}
}
\value{
ts: [[n]][nt, nroi] the ts loaded from the specified directory. list of n subjects, each with nt timesteps and nroi rois

dataset: [n] a vector of the dataset ids for each subject.

atlas: [n] a vector of the atlas ids for each subject.

subjects: [n] the subject ids

sessions: [n] the session ids

tasks: [n] the task ids

runs: [n] the run ids
}
\description{
A utility to open timeseries information from multiple datasets. Assumes data is formatted by the BIDs spec.
This means that derivatives are named as follows: sub-[###]_task-[abcde]_ses-[###]_run-[###]_(generic info).file
}
