% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2zsc.R
\name{fmriu.obs2zsc}
\alias{fmriu.obs2zsc}
\title{signal to Z-scores}
\usage{
fmriu.obs2zsc(signal, rtype = "list")
}
\arguments{
\item{signal:}{[[n]][nt, nroi] computes the z-scores of each of the nt observations with respect to the scaling of each roi.}

\item{rtype='list':}{the type of output to return. Options are 'list' and 'array'.}
}
\value{
zscore_data: [[n]][nt, nroi] the z-scored timeseries for each roi.
}
\description{
Scales observations by their z-scores for a given signal.
}
\author{
Eric Bridgeford
}
