% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{fmriu.plot.plot_graph}
\alias{fmriu.plot.plot_graph}
\title{Graph plot}
\usage{
fmriu.plot.plot_graph(mtx, title = "", xlabel = "ROI", ylabel = "ROI",
  legend = "metric", itype = "sq", fsize = 12, rem_diag = FALSE,
  limits = c(0, 1))
}
\arguments{
\item{mtx}{[nroi, nroi] or [nt, nroi]: the input. Can either be a square matrix of featuresxfeatures, or a timeseries of observationsxfeatures (see itype parameter).}

\item{title="":}{the title for the square plot.}

\item{xlabel="ROIs":}{the x label for the square plot.}

\item{ylabel="ROIs":}{the y label for the square plot.}

\item{legend="":}{the legend title for the square plot.}

\item{itype="sq":}{the shape of the input. If "sq", the plot will be generated as is for the input signal. If "ts", we will assume the input is observationsxfeatures, and will correlate the features first.}

\item{fsize=12:}{the default font size for the plot text.}

\item{include_diag=FALSE:}{whether to include the diagonal from the plot.}

\item{limits=c(0, }{1): the limits for the fill color in the heatmap of our matrix.}
}
\value{
sqplot : a plot of the square.
}
\description{
A function that plots a graph.
}
\author{
Eric Bridgeford
}
