% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{fmriu.plot.plot_timeseries}
\alias{fmriu.plot.plot_timeseries}
\title{Timeseries plot}
\usage{
fmriu.plot.plot_timeseries(ts, title = "", xlabel = "TRs",
  ylabel = "intensity", legend = "ROI", fsize = 12)
}
\arguments{
\item{ts:}{[nt,nroi] the signal, containing an array of nt observations for nroi rois.}

\item{title="":}{the title for the timeseries plot.}

\item{xlabel="TRs":}{the x label for the timeseries plot.}

\item{ylabel="intensity":}{the y label for the timeseries plot.}

\item{legend="ROIs":}{the legend title for the timeseries plot.}

\item{fsize=10:}{the default font size for the plot text.}
}
\value{
tsplot : a plot of the timeseries.
}
\description{
A function that plots a timeseries.
}
\author{
Eric Bridgeford
}
