% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_graphs.R
\name{fmriu.thresh_graphs}
\alias{fmriu.thresh_graphs}
\title{Graph Thresholding}
\usage{
fmriu.thresh_graphs(graphs, method = NaN, t = 0.5, binarize = FALSE)
}
\arguments{
\item{graphs:}{[[n]][nroi, nroi] graphs to threshold.}

\item{method=NaN:}{the method to use for thresholding the graphs.
\describe{
  \item{'cutoff'}{use a cutoff value. Sets all elements in the graph below this value to zero.}
  \item{'ptile'}{use a percentile. Calculates the cutoff for each list element given the appropriate percentile, and sets elements under this percentile to zero.}
   \item{NaN}{do not apply any thresholding to the graphs.}
}}

\item{t=0.5:}{the threshold to use for the thresholding method selected. Behavior depends on the method above chosen:
\describe{
  \item{method == 'cutoff'}{$t$ corresponds to the cutoff value. All values in the graphs less than $t$ will be set to zero.}
  \item{method == 'ptile'}{$t$ corresponds to the normalized  percentile between 0 and 1. The value of $t$ corresponds to the $100*t$ percentile.}
  \item{method == NaN}{$t$ does nothing.}
}}

\item{binarize=`FALSE`:}{whether to binarize the graphs after the thresholding has been applied.
\describe{
 \item{FALSE}{do not binarize the graphs.}
 \item{TRUE}{binarize the graphs by setting all nonzero values to 1, and leaving all zero values as 0.}
}}
}
\value{
tgraphs: [[n]][nroi, nroi] the thresholded graphs.
}
\description{
A utility function for easily applying various methods of thresholding to graphs.
}
\examples{
# I want to binarize my graphs without applying any thresholding
# setting all edges nonzero to connected (1) and zero edges to disconnected (0).
thresh_graphs(graphs, binarize=TRUE)

# For each graph, I want to threshold the botton 70\% of edges
# and then binarize such that the bottom 70\% of edges are disconnected (0)
# and the top 30\% of edges are connected (1).
thresh_graphs(graphs, method='ptile', t=0.5, binarize=TRUE)

# For each graph, remove all edges less than 0.5.
thresh_graphs(graphs, method='cutoff', t=0.5)

}
\author{
Eric Bridgeford
}
