% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2corr.R
\name{fmriu.time.obs2corr}
\alias{fmriu.time.obs2corr}
\title{Observations to Correlation
a utility to convert timeseries to correlation matrices.}
\usage{
fmriu.time.obs2corr(observations, rtype = "list")
}
\arguments{
\item{observations:}{[[subs]][timesteps, rois] a list of observations for a particular subject.}

\item{rtype='list':}{the type of output to return. Options are 'list' and 'array'.}
}
\value{
corr_data: [[subs]][rois, rois] the locally correlated roi timeseries.
}
\description{
Observations to Correlation
a utility to convert timeseries to correlation matrices.
}
\author{
Eric Bridgeford
}
