% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_graphs.R
\name{fmriu.io.collection.open_graphs}
\alias{fmriu.io.collection.open_graphs}
\title{Open Graphs Collection}
\usage{
fmriu.io.collection.open_graphs(basepath, datasets = "", atlases = "",
  gname = "connectomes", fmt = "graphml", verbose = FALSE,
  rtype = "list", flatten = FALSE)
}
\arguments{
\item{basepath:}{[1] a path to a properly spec'd directory. See the example in `fmriutils.html` for details.}

\item{datasets:}{[d] the datasets as a vector. Note that we expect a folder `<basepath>/<dataset>` for `dataset in datasets`.}

\item{atlases:}{[a] the atlases as a vector. Note that we expect a folder `<basepath>/<dataset>/<gname>/<atlas>` for `atlas in atlases`.}

\item{gname='connectomes':}{the folder name in which graphs are stored. Suggested options are ['connectomes', 'graphs'].}

\item{fmt="graphml"}{a parameter indicating the format for graphs to be read in as. Options are ['graphml', 'edgelist'].}

\item{verbose=FALSE:}{whether to print the id of the scan being loaded.}

\item{rtype='list':}{the type of output to return. Options are 'list' and 'array'. Note that if rtype is 'array', all of the graphs must be derived from parcellations with the same number of vertices.}

\item{flatten=FALSE}{a parameter to flatten the array if the rtype is set to 'array'.
\describe{
  \item{TRUE}{If rtype == 'array', then returns an [n x r^2] array}
  \item{FALSE}{If rtype == 'array', then returns an [n x r x r] array}
}}
}
\value{
graphs: [[n]][nroi, nroi] the graphs loaded from the specified directory. Shape depends on the parameters specified.

dataset: [n] a vector of the dataset ids for each subject.

atlas: [n] a vector of the atlas ids for each subject.

subjects: [n] the subject ids

sessions: [n] the session ids

tasks: [n] the task ids

runs: [n] the run ids
}
\description{
A utility to open graphs information from multiple datasets. Assumes data is formatted by the BIDs spec.
This means that derivatives are named as follows: sub-[###]_task-[abcde]_ses-[###]_run-[###]_(generic info).file
}
