% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_graphs.R
\name{fmriu.io.open_graphs}
\alias{fmriu.io.open_graphs}
\title{Open Graphs}
\usage{
fmriu.io.open_graphs(fnames, dataset_id = "", atlas_id = "",
  fmt = "graphml", verbose = FALSE, rtype = "list")
}
\arguments{
\item{fnames:}{[n] a vector of filenames, with separation by underscores IE, dataset_subject_run_(other information).rds}

\item{dataset_id="":}{[1] the dataset id in the filenames.}

\item{atlas_id="":}{[1] the atlas id in the filenames.}

\item{fmt="graphml"}{: a parameter idicating the format for graphs to be read in as. Options are ['graphml'].}

\item{verbose=FALSE}{: whether to print the iteration being loaded.}

\item{rtype='list':}{the type of output to return. Options are 'list' and 'array'.}
}
\value{
gr: [[subs]][nt, nroi] the gr loaded from the specified file names. list of n subjects, each with nt timesteps and nroi rois

dataset: [n] a vector of the dataset ids for each subject.

subjects: [n] the subject ids

sessions: [n] the run ids
}
\description{
A utility to open graph information from a dataset.Assumes data is formatted by the BIDs spec.
This means that derivatives are named as follows: sub-[#####]_task-[abcde]_ses-[#####]_(generic info).file
}
