% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2zsc.R
\name{fmriu.obs2zsc}
\alias{fmriu.obs2zsc}
\title{signal to Z-scores}
\usage{
fmriu.obs2zsc(signal, scale = "roi")
}
\arguments{
\item{signal:}{[[n]][nt, nroi] timeseries or [[n]][nroi, nroi] graphs to z-score.}

\item{scale='roi':}{how to handle scaling.
+ 'roi' columns of each element of our input (ROIs; our features, this gives a z-score per ROI; suggested if inputs are timeseries).
+ 'global' by the overall mean (this gives a global z-scorel; suggested for graphs).}
}
\value{
zscore_data: [[n]][nt, nroi] timeseries or [[n]][nroi, nroi] graphs the z-scored data.
}
\description{
Scales observations by their z-scores for a given signal.
}
\author{
Eric Bridgeford
}
