% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dataseries.R
\name{fmriu.plot.plot_dataseries}
\alias{fmriu.plot.plot_dataseries}
\title{Dataseries plot}
\usage{
fmriu.plot.plot_dataseries(ds, xax = NaN, title = "", xlabel = "TRs",
  ylabel = "intensity", legend = "ROI", fsize = 12)
}
\arguments{
\item{ds:}{[nt,nroi] the signal, containing an array of nt observations for nroi rois.}

\item{title="":}{the title for the dataseries plot.}

\item{xlabel="TRs":}{the x label for the dataseries plot.}

\item{ylabel="intensity":}{the y label for the dataseries plot.}

\item{legend="ROIs":}{the legend title for the dataseries plot.}

\item{fsize=10:}{the default font size for the plot text.}
}
\value{
dsplot : a plot of the dataseries.
}
\description{
A function that plots a dataseries, such as a time series or frequency series.
}
\author{
Eric Bridgeford
}
